refactor(internals): Reorganise internals code
